refactor(vtscore): split config.py into a package along its six seams
`vtscore/config.py` was 933 lines holding six unrelated things — filesystem
roots, scalar tunables, HF model ids, CUDA/precision resolution, transformers
image-processor negotiation, and `CoreConfig` — sharing nothing but the module
name, in the single most-imported module in the library.

Split into `vtscore/config/`, layered so each submodule reads only from the ones
above it: `paths` → `runtime` → `models` → `device` → `processor_backend` →
`core_config`. Every top-level statement moved verbatim; the only two edits to
moved code are `_REPO_ROOT` (now `parents[2]`, since `paths.py` sits one level
deeper) and `image_processor_call_kwargs`, which calls through the `device`
*module* so one stub target covers both it and the precision resolvers.

The import surface is unchanged: `__init__` re-exports all 74 public names and
lists them in `__all__`, so `vtscore.config.X` and `from vtscore.config import X`
resolve exactly as before. Private names are deliberately not re-exported — a
stub installed on a copy of `_cuda_can_run` or `_core_config_builder` would be
silently ignored, so the copy should not exist.

Two consequences, both test-only, are documented in the package docstring and in
`vtscore/docs/packages/config.md`:

- A stub goes where the *reader* is. Callers outside the package resolve
  `vtscore.config.X`, so patching the package still reaches them; a function
  inside a submodule resolves its own module global. The affected patch targets
  (`allocated_cpus`, `_cuda_can_run`, `_transformers_major`,
  `_core_config_builder`, the precision resolvers' `resolve_device`) are
  retargeted at the submodule that owns them. One of these was masked by a
  `raising=False` in `test_docs_quickstart`'s `detectors_dir` fixture; that is
  now a hard target and the `raising=False` is gone.
- `importlib.reload(vtscore.config)` no longer re-reads the environment, because
  the submodules are cached. `config._reload_all()` reloads them in dependency
  order and then the package; the ~20 env-var tests use it, and their restore
  fixtures now snapshot the whole package tree rather than just the package.

`tests_lib/core/test_config_package.py` pins the three things a seventh
submodule could break silently: `_RELOAD_ORDER` covers every file, that order is
a real dependency order (checked against the submodules' own imports), and no
public submodule name is missing from `__all__`.

Also: `scripts/check-vtscore-docs.py`'s `DOC_FOR` key follows the file to a
directory; `BEATS_EMBED_DIM` leaves `.vulture-whitelist.py` (the `__all__` entry
is a real reference now, so the whitelist line suppressed nothing and the gate
said so); the config doc grows a package-layout section and the six env vars its
"every env var, in one place" table was missing.

Closes #3375
